alx_travel_app/requirement.txt
django
djangorestframework
django-cors-headers
drf-yasg
celery
rabbitmq
django-environ
mysqlclient
packaging
setuptools
